repositoryファイルの型もLongに変更

+++ backend/src/main/java/chibainfo5/es_manager/repository/EsRepository.java
@@ -1,7 +1,7 @@
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface EsRepository extends JpaRepository<EsEntity, Integer> {
+public interface EsRepository extends JpaRepository<EsEntity, Long> {
     // 必要なメソッドを追加予定（勉強中）
-    List<EsEntity> findByFirstName(Int esId);
+    List<EsEntity> findByFirstName(Long esId);
 }
